Modified the scope to use Google profile and email scope instead of the Google Plus scope
Check this for reference :
https://developers.google.com/+/web/api/rest/oauth
In some case, the user won't have a Google Plus profile and the
profile.displayName will be empty, which is not really nice if you need
to do ACL or remove access to a user. Using the email seems more
appropriate since a user can change it's display name, and by doing so,
it will create a new XO account.